refactor: update module imports
